integrated order-service with inventory-service for order creation

+++ order-service/src/main/java/demo/repository/OrderRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import demo.entity.Order;
 
-public interface OrderRepository extends JpaRepository<Order,Integer> 
+public interface OrderRepository extends JpaRepository<Order,Long>
 {
 
 }
